requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
convertdate==2.3.2
Django==3.2.7
django-db==0.0.7
django-reset-migrations==0.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
geographiclib==1.52
geoip2==2.9.0
geolocation-python==0.2.2
geopy==2.2.0
hijri-converter==2.2.2
holidays==0.11.2
idna==3.2
importlib-metadata==4.8.1
IP2Location==8.6.2
kombu==5.1.0
korean-lunar-calendar==0.2.1
maxminddb==2.2.0
prompt-toolkit==3.0.20
PyJWT==1.7.1
PyMeeus==0.5.11
PyMySQL==1.0.2
python-dateutil==2.8.2
python-geoip==1.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.1